feat: adds cache to avoid repeated calls in loadUserSpaceData
- Implemented a cache mechanism in the  function to avoid repeated calls with the same parameters.
- Reduces unnecessary requests and improves performance.
- Includes logs to identify cache hits.
